requirements.txt
# KEXP Display Requirements

# KEXP API access
requests>=2.31.0

# Image processing
Pillow>=10.0.0

# RGB Matrix library (install separately - see README)
# Follow instructions at: https://github.com/hzeller/rpi-rgb-led-matrix/tree/master/bindings/python
